// src/app/models/admin-models/model.reversFormBig.ts
export class ReversFormBig
{
    id: number;
    
    is_review?:number;

    name?: string;
    phone?: string;
    email?: string;
    theme?: string;
    message?: string;
    country?: string;
    street?: string;
    company?: string;
    city?: string;
    post?: number;
    
    createdAt?: string;
    updatedAt?: string;
}